Feat: flesh out observe module
Add all required buttons and fields. Some buttons
are exclusive to each other, such as all image
type buttons and the start, stop, and resolve
buttons. Some buttons have tooltips, and more will
likely be added later. All fields update a value
in the object to be sent to the backend.