refactor: move registration to dedicated protected route
Our plan for the moment is to place the
registration functionality, for registering new
users, behind authentication protection so that
only registered users may create new ones,
limiting access to the tool to those who are
permitted by an admin user.

This commit also initializes the usage of an
email and observer name for the system, instead
of only a username. In the future, we will use
the observer name as the public facing name given
on FITS files, and the email will be used as the
authenticating identification.