fix(frontend): rename "observation_type" → "obs_type"
The database currently uses "obs_type" to hold
onto the observation type. The frontend passes
this value around as "observation_type" when it
would be more convenient to use the name the
database has given it. It would be more convenient
because, with the other way, we must translate.

A future improvement should be to make all column
names that have an equivalent name in the headers
comply with the established header form, but
lowercase. For example, FITS headers use
"OBSTYPE", meaning the database should have
"obstype".