feat: dynamically update Log progress
The application should display the progress of
queued observations. This commit adds the ability
to do that using three stages: "Pending", "In
Progress", and "Complete", and displays these
stages using MUI chips.

The Pending and Complete statuses can be
determined using information from the
observation table, but this table does not track
when a specific observation has been started. The
only means we currently have to do this is via the
status table, which the camera updates constantly
as it works. To get the progress info we want for
the Logsheet component, we must currently go
through the status table.