refactor(App): improve authentication overlay
We use an overlay on pages that require
authentication; as the backend authenticates the
user, we display a background that indicates
that work is being done. This overlay currently
uses a dark color and displays a progress spinner
immediately. The dark color is particularly
annoying and displaying the spinner immediately is
bad practice.

This commit makes the background transparent and
loads the spinner after 80ms.