refactor(App): improve authentication overlay
We use a MUI overlay to indicate when a page is
currently loading due to authentication. This
overlay originally used a dark background and
start displaying a circular progress notifier
immediately. The dark background made navigation
look jittery and the overlay feels too immediate,
which MUI warns of in its documentation

Make the background invisible and start the
progress notifier after a small period of time.