refactor(user management): init new row on backend first
With the initial implementation of the user
management table's communication with the
database, I instructed the table to create an
empty row on the frontend first, then commit it
to the db after it had been saved. This created
issues with managing which items had legitimate
IDs, given that it managed both an ID to the data
grid and the database.

This commit inits an empty database row on
creation of a new row in the table. As a
consequence, this fixes a known issue where
editing an existing row duplicated the row in the
database and table.

Other changes:
- Fixes str validation for the username