fix(Observe): cannot use non-object obs type without populating object field
Problem: The observe form does not allow the
user to request a non-object observation type
if the object text input does not have content

Solution: Use conditional validation to only
require that the object field has content when
observation type is set to something other than
object.

Also, add catches for when the number of exposures
and the exposure duration are made empty and
submitted. Before this, that issue would be caught
but the error message was ugly. The catch is only
to add nice error messages