fix(spectrograph): update how observation type is sent to spectrograph
The spectrograph required a sleep to make sure
that it adjusted its observation type before the
camera started working, so that the spectrograph
was already prepared for what the camera has to
do. #111 updated the observation API to set the
observation type and wait for successful
confirmation before permitting the camera to begin
working. This commit fixes how data is sent to the
spectrograph when the observation type update
request is made.

Co-authored-by: koseward <kos34@nau.edu>