fix(observation form): persist access to form
Problem: Access to the observation form does not
persist across page navigation or log out. For
example, if the user starts an observation and
changes pages, they can no longer access the
"Stop" button to terminate the observation.

Solution: When navigating to the observation
request page, check if the system is currently
busy using a new "System" status. If not, allow
all users access to the form. If it is busy, check
if the current user is the owner of the current
observation. If so, allow them access to the
"Stop" button. If not, deny access to the form
submission and termination buttons.