feat(api/main/__init__): configure app from objects
In preparation for building our integration tests
between the app and client instruments, it is
necessary to allow the developers to configure the
application's settings, specifically the SQLAlchemy
database URI, for testing. This allows for testing
with a database other than that used for
production (and allows you to have a different
database URI for development as well).

To make this possible, this commit moves
configuration away from just an .env file and
incorporates configuring the application from
Python objects, as described in Flask's
documentation (see the link below). Now, the .env
file simply indicates to the application init what
configuration object to use, and the objects are
configured in a separate file.

Docs: https://flask.palletsprojects.com/en/2.3.x/config/#development-production