build(docker-compose): connect API container to /tmp
The cameras are currently written to write their
output to a file in the host file system's temporary
directory. To handle the contents of this file,
the application back-end must have access to where
the file is written. This commit mounts the host's
/tmp dir to the container's /tmp dir, allowing the
API to access the location where the camera writes
its output.

Note: this relies upon the OS's temporary dir
being named "/tmp". Another naming scheme will
break this. The camera (and probably the
spectrograph) should be containerized for maximum
portability.