fix: map FITS output dir from container to host
During containerization, I forgot to provide a
means for the host to access the output location
the file writer uses for FITS files (configured
in app_config.py). This commit changes that output
location so that it is always at the root of the
repository, whether run in or outside of a
container. This choice was made because that is
currently where our database data is written as
well and is a safe location on the host machine.